refactor: [#157] extract API contexts
