+++ src/apis/context/mod.rs
@@ -1,4 +1,4 @@
 pub mod auth_key;
-pub mod peer;
 pub mod stats;
 pub mod torrent;
+pub mod whitelist;
