+++ src/apis/context/stats/responses.rs
@@ -1,0 +1,8 @@
+use axum::response::Json;
+
+use super::resources::Stats;
+use crate::tracker::services::statistics::TrackerMetrics;
+
+pub fn stats_response(tracker_metrics: TrackerMetrics) -> Json<Stats> {
+    Json(Stats::from(tracker_metrics))
+}
