+++ src/apis/context/torrent/routes.rs
@@ -1,0 +1,17 @@
+use std::sync::Arc;
+
+use axum::routing::get;
+use axum::Router;
+
+use super::handlers::{get_torrent_handler, get_torrents_handler};
+use crate::tracker::Tracker;
+
+pub fn add(router: Router, tracker: Arc<Tracker>) -> Router {
+    // Torrents
+    router
+        .route(
+            "/api/torrent/:info_hash",
+            get(get_torrent_handler).with_state(tracker.clone()),
+        )
+        .route("/api/torrents", get(get_torrents_handler).with_state(tracker))
+}
