+++ src/apis/mod.rs
@@ -1,6 +1,10 @@
-pub mod handlers;
+pub mod context;
 pub mod middlewares;
-pub mod resources;
 pub mod responses;
 pub mod routes;
 pub mod server;
+
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct InfoHashParam(pub String);
