+++ tests/tracker_api.rs
@@ -132,7 +132,7 @@ mod tracker_apis {
     mod for_stats_resources {
         use std::str::FromStr;
 
-        use torrust_tracker::apis::resources::stats::Stats;
+        use torrust_tracker::apis::context::stats::resources::Stats;
         use torrust_tracker::protocol::info_hash::InfoHash;
         use torrust_tracker_test_helpers::configuration;
 
@@ -206,8 +206,8 @@ mod tracker_apis {
     mod for_torrent_resources {
         use std::str::FromStr;
 
-        use torrust_tracker::apis::resources::torrent::Torrent;
-        use torrust_tracker::apis::resources::{self, torrent};
+        use torrust_tracker::apis::context::torrent::resources::peer::Peer;
+        use torrust_tracker::apis::context::torrent::resources::torrent::{self, Torrent};
         use torrust_tracker::protocol::info_hash::InfoHash;
         use torrust_tracker_test_helpers::configuration;
 
@@ -383,7 +383,7 @@ mod tracker_apis {
                     seeders: 1,
                     completed: 0,
                     leechers: 0,
-                    peers: Some(vec![resources::peer::Peer::from(peer)]),
+                    peers: Some(vec![Peer::from(peer)]),
                 },
             )
             .await;
